Update README.md
Removed the "Hosting Yourself" section. There are plenty of ways to host a bot, and hosting should not be in the scope of this rep. There are already more thorough guides for this, on the web. Added "Permissions" section, which should be important to know, if anyone wants to self host. Updated "Commands" with the new commands