bugfix
Previously only a single boolean "is_uwu" controlled if "uwu_all" was currently toggled on and there were no checks, on the channel the message (to uwuify) was coming from. This caused the "uwu_all" command to be global. In other words, toggling it on, toggled it on for every single channel the bot was in. I fixed this by instead having an array, with all the channels the command is active in